+++ test/sandbox/test/rendering_test.rb
@@ -1075,8 +1075,8 @@ class RenderingTest < ViewComponent::TestCase
   end
 
   def test_multiple_inline_renders_of_the_same_component
-    component = ErbComponent.new(message: "foo")
-    render_inline(InlineRenderComponent.new(items: [component, component]))
+    items = [ErbComponent.new(message: "foo"), ErbComponent.new(message: "foo")]
+    render_inline(InlineRenderComponent.new(items: items))
     assert_selector("div", text: "foo", count: 2)
   end
 
@@ -1375,6 +1375,18 @@ class RenderingTest < ViewComponent::TestCase
     assert_equal(1, Instrumenter.count)
 
     assert_text("Hi!")
+  end
+
+  def test_around_render_html_unsafe_output_is_escaped
+    warnings = capture_warnings { render_inline(UnsafeAroundRenderComponent.new) }
+
+    assert_includes @rendered_content, "&lt;script&gt;alert(1)&lt;/script&gt;"
+    refute_includes @rendered_content, "<script>"
+    assert_predicate @rendered_content, :html_safe?
+    assert(
+      warnings.any? { |warning| warning.include?("around_render returned an HTML-unsafe string") },
+      "Rendering UnsafeAroundRenderComponent did not emit an HTML safety warning"
+    )
   end
 
   def test_render_partial_with_yield